重构了项目结构，把reducer和saga都做了拆分， 不在封装actionCreators了，api也做了极简的优化，不在搞这么复杂了。并且自己简单的封了一下路由
