+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^0.19.2",
     "immer": "^7.0.7",
     "marked": "^1.1.1",
+    "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-loadable": "^5.5.0",
