+++ public/index.html
@@ -7,7 +7,7 @@
   <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
   <link rel="stylesheet" href="%PUBLIC_URL%/css/main.css">
   <link rel="stylesheet" href="%PUBLIC_URL%/css/css.css">
-  <link href="//code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css" rel="stylesheet" type="text/css">
+  <!-- <link href="//code.ionicframework.com/ionicons/2.0.1/css/ionicons.min.css" rel="stylesheet" type="text/css"> -->
 
   <!--
       Notice the use of %PUBLIC_URL% in the tag above.
