+++ src/constants/actionTypes.js
@@ -1,0 +1,1 @@
+export const CHANGE_COMMON_NAV = "change_common_nav";
