+++ src/constants/confog.js
@@ -1,2 +1,1 @@
 export const API_URL = "https://conduit.productionready.io/api";
-export default API_URL;
