+++ src/pages/Test.js
@@ -1,11 +1,20 @@
 import React, { PureComponent } from "react";
+import history from "../utils/history";
 class Test extends PureComponent {
   constructor(props) {
     super(props);
     this.state = {};
   }
   render() {
-    return <div>index</div>;
+    return (
+      <div
+        onClick={() => {
+          history.push("/");
+        }}
+      >
+        test
+      </div>
+    );
   }
 }
