+++ src/saga/auth.js
@@ -1,0 +1,8 @@
+import { put, takeEvery } from "redux-saga/effects";
+import { saveToken, destroyToken } from "../utils";
+import history from "../utils/history";
+function* authSaga() {
+
+}
+
+export default authSaga;
