+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from "redux";
 import createSagaMiddleware from "redux-saga";
 import reducer from "./reducer";
-import saga from "./saga";
+// import saga from "./saga";
 
 // create the saga middleware
 const sagaMiddleware = createSagaMiddleware();
@@ -13,4 +13,4 @@ const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
 const enhancer = composeEnhancers(applyMiddleware(sagaMiddleware));
 export default createStore(reducer, enhancer);
 
-sagaMiddleware.run(saga);
+// sagaMiddleware.run(saga);
