+++ src/store/reducer.js
@@ -1,123 +1,10 @@
-import produce from "immer";
-import { saveToken } from "../utils";
-import {
-  INITUSERINFO,
-  USERERROR,
-  CHANGEAPPNAV,
-  INITALLTAG,
-  CHANGEHOMETAB,
-  INITGLOBALARTICLESlIST,
-  CHANGEGLOBALOFFSET,
-  INITARTICLEDETAILS,
-  UPDATEARTICLEAUTHOR,
-  INITARTICLECOMMENTS,
-  ADDARTICLECOMMENTS,
-  CHANGE_CUTTENT_USER_TEXT,
-  CLEAN_USER_INFO,
-  CLEAN_EDITOR,
-  SET_EDITOR,
-  ADD_EDITOR_TAG,
-  DELETE_EDITOR_TAG,
-  INIT_PROFILES_USER,
-} from "./actionTypes";
+import { combineReducers } from "redux";
+import common from "../reducer/common";
+import auth from "../reducer/auth";
+import home from "../reducer/home";
 
-const defaultState = {
-  userInfo: {},
-  currentUser: {},
-  navActive: "Home",
-  tags: [],
-  tab: "Global",
-  article: [],
-  offset: 0,
-  limit: 10,
-  articlesCount: 0,
-  errors: null,
-  articleDetails: "",
-  currentArticleComment: [],
-  currentEditor: {
-    title: "",
-    body: "",
-    description: "",
-    tagList: [],
-    tagValue: "",
-  },
-  currentProfiles: {
-    bio: "",
-    following: false,
-    image: "",
-    username: "",
-  },
-};
-
-export default produce((state = defaultState, action) => {
-  switch (action.type) {
-    case INITUSERINFO:
-      state.userInfo = action.data;
-      state.currentUser = action.data;
-      saveToken(action.data.token);
-      break;
-    case USERERROR:
-      state.errors = action.data;
-      break;
-    case CHANGEAPPNAV:
-      state.navActive = action.data;
-      break;
-    case INITALLTAG:
-      state.tags = action.data;
-      break;
-    case CHANGEHOMETAB:
-      state.tab = action.data;
-      break;
-    case INITGLOBALARTICLESlIST:
-      state.article = action.data.articles;
-      state.articlesCount = action.data.articlesCount;
-      break;
-    case CHANGEGLOBALOFFSET:
-      state.offset = action.data;
-      break;
-    case INITARTICLEDETAILS:
-      state.articleDetails = action.data;
-      state.currentEditor = action.data;
-      break;
-    case UPDATEARTICLEAUTHOR:
-      state.articleDetails.author = action.data;
-      break;
-    case INITARTICLECOMMENTS:
-      state.currentArticleComment = action.data;
-      break;
-    case ADDARTICLECOMMENTS:
-      state.currentArticleComment.unshift(action.data);
-      break;
-    case CHANGE_CUTTENT_USER_TEXT:
-      state.currentUser = action.data;
-      break;
-    case CLEAN_USER_INFO:
-      state.userInfo = {};
-      state.currentUser = {};
-      break;
-    case CLEAN_EDITOR:
-      state.currentEditor = {
-        title: "",
-        body: "",
-        description: "",
-        tagList: [],
-        tagValue: "",
-      };
-      break;
-    case SET_EDITOR:
-      state.currentEditor = action.data;
-      break;
-    case ADD_EDITOR_TAG:
-      state.currentEditor.tagList.push(action.data);
-      state.currentEditor.tagValue = "";
-      break;
-    case DELETE_EDITOR_TAG:
-      state.currentEditor.tagList.splice(action.data, 1);
-      break;
-    case INIT_PROFILES_USER:
-      state.currentProfiles = action.data;
-      break;
-    default:
-      return state;
-  }
+export default combineReducers({
+  common,
+  auth,
+  home,
 });
