+++ src/utils/history.js
@@ -1,6 +1,2 @@
-// import createHistory from 'history/createBrowserHistory';
-
-// export default createHistory();
-
 import { createHashHistory } from "history";
 export default createHashHistory();
